requirements.txt
accelerate==1.7.0
aiofiles==25.1.0
aiohttp==3.12.13
aiohttp==3.12.12
apex==0.9.10dev
bert_score==0.3.13
cachetools==5.5.2
clearml==2.0.2
codetiming==1.4.0
cupy_cuda12x==13.4.1
datasets==3.5.0
dill==0.3.8
einops==0.8.1
faiss==1.5.3
fastapi==0.121.2
fastmcp==2.13.1
filelock==3.20.0
flash_attn==2.7.3
huggingface_hub==0.35.3
hydra-core==1.3.2
keybert==0.9.0
langchain_core==1.0.5
langgraph==1.0.3
latex2sympy2_extended==1.10.2
liger_kernel==0.6.3
mathruler==0.1.0
mcp==1.21.2
mlflow==3.6.0
modelscope==1.29.0
msgspec==0.19.0
nltk==3.9.1
numpy==1.25.0
nvtx==0.2.13
omegaconf==2.3.0
openai==2.8.1
packaging==25.0
pandas==1.5.3
peft==0.17.0
Pillow==12.0.0
pip==25.1
psutil==7.1.1
psutil==7.0.0
pydantic==2.12.4
pyext==0.7
pylatexenc==2.10
pyserini==1.3.0
pytest==9.0.1
pyzmq==27.1.0
qwen_vl_utils==0.0.14
regex==2024.11.6
Requests==2.32.5
rich==14.2.0
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.6.2
scikit_learn==1.7.2
sentence_transformers==5.1.0
setuptools==80.9.0
starlette==0.50.0
swanlab==0.7.2
sympy==1.13.1
tensordict==0.6.2
textual==6.6.0
torch_npu==2.8.0
torchdata==0.11.0
torchvision==0.21.0
tqdm==4.67.1
trackio==0.9.1
triton==3.2.0
trl==0.9.6
typer==0.20.0
ujson==5.11.0
uvicorn==0.38.0
wandb==0.23.0
weave==0.52.17
